+++ test/Makefile
@@ -2,10 +2,21 @@
 CC		:= gcc
 CFLAGS	:= -Wall -Wextra -Werror -g -fsanitize=address
 
-all: mmap
+SRCDIR	:= ./
+OBJDIR	:= ./
+SRCS	:=\
+			yo_malloc.c\
+			yo_list.c\
+			yo_allocation.c\
+
+OBJS	:= $(SRCS:$(SRCDIR)%.c=$(OBJDIR)%.o)
+
+all: malloc
 
 
-.PHONY:	mmap
-mmap:
-	$(CC) $(CFLAGS) -o $@ $@.c
+.PHONY:	malloc
+malloc:	$(OBJS)
+	$(CC) $(CFLAGS) -o $@ main.c $(OBJS)
 	./$@
+
+
